refactor: Event Dto 위치 수정
