+++ src/main/java/com/happiday/Happi_Day/domain/controller/EventController.java
@@ -1,8 +1,8 @@
 package com.happiday.Happi_Day.domain.controller;
 
-import com.happiday.Happi_Day.domain.dto.event.EventCreateDto;
-import com.happiday.Happi_Day.domain.dto.event.EventResponseDto;
-import com.happiday.Happi_Day.domain.dto.event.EventUpdateDto;
+import com.happiday.Happi_Day.domain.entity.event.dto.EventCreateDto;
+import com.happiday.Happi_Day.domain.entity.event.dto.EventResponseDto;
+import com.happiday.Happi_Day.domain.entity.event.dto.EventUpdateDto;
 import com.happiday.Happi_Day.domain.service.EventService;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
