+++ src/main/java/com/happiday/Happi_Day/domain/entity/event/dto/EventResponseDto.java
@@ -1,10 +1,7 @@
-package com.happiday.Happi_Day.domain.dto.event;
+package com.happiday.Happi_Day.domain.entity.event.dto;
 
 import com.happiday.Happi_Day.domain.entity.event.Event;
-import jakarta.validation.constraints.NotBlank;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.Setter;
+import lombok.*;
 
 import java.time.LocalDateTime;
 import java.util.List;
@@ -12,7 +9,8 @@ import java.util.List;
 @Getter
 @Setter
 @Builder
-public class EventListResponseDto {
+public class EventResponseDto {
+
     private Long id;
 
     private String username;
@@ -22,22 +20,25 @@ public class EventListResponseDto {
     private LocalDateTime startTime;
 
     private LocalDateTime endTime;
-    
+
+    private String description;
+
     private String location;
 
-    private String thumbnailUrl;
+    private String imageUrl;
 
     private List<String> artists;
 
-    public static EventListResponseDto fromEntity(Event event) {
-        return EventListResponseDto.builder()
+    public static EventResponseDto fromEntity(Event event) {
+        return EventResponseDto.builder()
                 .id(event.getId())
                 .username(event.getUser().getNickName())
                 .title(event.getTitle())
                 .startTime(event.getStartTime())
                 .endTime(event.getEndTime())
+                .description(event.getDescription())
                 .location(event.getLocation())
-                .thumbnailUrl(event.getThumbnailUrl())
+                .imageUrl(event.getImageUrl())
 //                .artists(event.getArtists()) TODO 아티스트 네임 추가
                 .build();
     }
